requirements.txt
discord.py>=2.3.0
wavelink>=3.2.0
python-dotenv>=1.0.0
